--- src/components/UI/Icons/container-close.tsx
@@ -1,16 +1,0 @@
-export default function ContainerCloseIcon() {
-  return (
-    <svg
-      width="14"
-      height="14"
-      viewBox="0 0 14 14"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M1.40942 0.490063C1.15558 0.236222 0.744026 0.236222 0.490185 0.490063C0.236344 0.743904 0.236344 1.15546 0.490185 1.4093L6.0808 6.99991L0.590648 12.4901C0.336807 12.7439 0.336808 13.1555 0.590649 13.4093C0.844489 13.6631 1.25605 13.6631 1.50989 13.4093L7.00004 7.91915L12.4902 13.4093C12.744 13.6631 13.1556 13.6631 13.4094 13.4093C13.6633 13.1555 13.6633 12.7439 13.4094 12.4901L7.91928 6.99991L13.5099 1.4093C13.7637 1.15546 13.7637 0.743904 13.5099 0.490063C13.256 0.236222 12.8445 0.236222 12.5906 0.490063L7.00004 6.08068L1.40942 0.490063Z"
-        fill="#505050"
-      />
-    </svg>
-  );
-}
