+++ src/components/preloader/preloader.module.scss
@@ -1,0 +1,20 @@
+.preloader__wrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  .preloader {
+    animation: rotate 1s infinite;
+  }
+}
+
+@keyframes rotate {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
