+++ src/pages/map-location/map-location.tsx
@@ -1,9 +1,9 @@
 import styles from "./map-location.module.scss";
 import Container from "components/container/container";
-import PreloaderIcon from "components/UI/Icons/preloader";
 import { YMaps, Map, Placemark } from "@pbe/react-yandex-maps";
 import { useState } from "react";
 import { location } from "utils/user-info/location";
+import Preloader from "components/preloader/preloader";
 
 export default function MapPage() {
   const [mark, setMark] = useState<number[]>();
@@ -12,18 +12,10 @@ export default function MapPage() {
     setMark(location.mark);
   };
 
-  const preloader = (
-    <div className={styles.preloader__wrapper}>
-      <div className={styles.preloader}>
-        <PreloaderIcon />
-      </div>
-    </div>
-  );
-
   return (
     <Container title="Basic map">
       <div className={styles.map__wrapper}>
-        {!mark && preloader}
+        {!mark && <Preloader />}
         <YMaps>
           <Map
             className={styles.map}
